Code Cleanup, rename outputbc to execbc, using buffered Writer
